fixed and verified calcStPose, fixed calcShCenter, added shear_flow.py
